# backend/src/modules/blog/blog.routes.ts
import type { FastifyInstance } from "fastify";
import type { ZodTypeProvider } from "fastify-type-provider-zod";
import type { BlogPost } from "@prisma/client";
import { Prisma } from "@prisma/client";
import { z } from "zod";
import { authenticate } from "../../middleware/authenticate";
import { requireSiteRole } from "../../middleware/site-rbac";
import { ok } from "../../lib/envelope";
import { ApiSuccessSchema, CursorQuerySchema } from "../../schemas/common";
import { BlogCategorySchema, BlogPostSchema, ContentRevisionSchema, ContentRevisionSummarySchema } from "../../schemas/entities";
import { toBlogCategoryDto, toBlogPostDto, toContentRevisionDto, toContentRevisionSummaryDto } from "../../mappers";
import { NotFoundError } from "../../lib/errors";
import { parseCursor, encodeCursor, buildPageMeta } from "../../lib/pagination";
import { slugify } from "../../lib/slug";
import { startOfUtcDay } from "../../lib/date";
import { detectDeviceType } from "../../lib/device";
import { detectCountry } from "../../lib/geo";
import { touchVisitor } from "../../lib/live-visitors";
import { snapshotBeforeUpdate } from "../../lib/content-revisions";
import {
  CategoryIdParamSchema,
  CreateBlogCategoryRequestSchema,
  CreateBlogPostRequestSchema,
  LocaleQuerySchema,
  PostIdParamSchema,
  PostRevisionIdParamSchema,
  PostSlugParamSchema,
  UpdateBlogCategoryRequestSchema,
  UpdateBlogPostRequestSchema,
} from "./blog.schemas";

const WITH_CATEGORY = { category: true } as const;

/** Güncellemeden HEMEN ÖNCEKİ alan setini döner (bkz. ARCHITECTURE.md §10.1). */
function toBlogPostSnapshot(post: BlogPost): Record<string, unknown> {
  return {
    title: post.title,
    slug: post.slug,
    excerpt: post.excerpt,
    contentHtml: post.contentHtml,
    coverImageUrl: post.coverImageUrl,
    categoryId: post.categoryId,
    seoTitle: post.seoTitle,
    seoDescription: post.seoDescription,
    ogTitle: post.ogTitle,
    ogImageUrl: post.ogImageUrl,
    canonicalUrl: post.canonicalUrl,
    noIndex: post.noIndex,
    translations: post.translations,
  };
}

/**
 * §10.5 Çoklu Dil & Yerelleştirme — `locale=EN` verildiğinde `translations.EN`'deki
 * alan bazlı override'ları uygular (eksik alan TR/kanonik kolondan gelir).
 */
function applyLocale<T extends BlogPost>(post: T, locale?: "EN"): T {
  if (locale !== "EN") return post;
  const translations = (post.translations as Record<string, Record<string, unknown>> | null) ?? {};
  const en = translations.EN;
  if (!en) return post;

  return {
    ...post,
    title: typeof en.title === "string" ? en.title : post.title,
    seoTitle: typeof en.seoTitle === "string" ? en.seoTitle : post.seoTitle,
    seoDescription: typeof en.seoDescription === "string" ? en.seoDescription : post.seoDescription,
    ogTitle: typeof en.ogTitle === "string" ? en.ogTitle : post.ogTitle,
    canonicalUrl: typeof en.canonicalUrl === "string" ? en.canonicalUrl : post.canonicalUrl,
    excerpt: typeof en.excerpt === "string" ? en.excerpt : post.excerpt,
    contentHtml: typeof en.contentHtml === "string" ? en.contentHtml : post.contentHtml,
  };
}

/** `/admin/blog` prefix'i altında bağlanır (bkz. app.ts) — tüm durumlar (taslak dahil), authenticated. */
export async function adminBlogPostsRoutes(app: FastifyInstance) {
  const server = app.withTypeProvider<ZodTypeProvider>();
  server.addHook("preHandler", authenticate);

  server.get(
    "/",
    { schema: { querystring: CursorQuerySchema, response: { 200: ApiSuccessSchema(z.array(BlogPostSchema)) } } },
    async (request, reply) => {
      const { cursor, limit } = request.query;
      const cursorSeq = parseCursor(cursor);

      const rows = await app.prisma.blogPost.findMany({
        where: cursorSeq ? { seq: { gt: cursorSeq } } : {},
        orderBy: { seq: "asc" },
        take: limit,
        include: WITH_CATEGORY,
      });

      return reply.send(ok(rows.map(toBlogPostDto), buildPageMeta(rows, limit)));
    }
  );

  server.post(
    "/",
    {
      preHandler: requireSiteRole("ADMIN", "EDITOR"),
      schema: { body: CreateBlogPostRequestSchema, response: { 201: ApiSuccessSchema(BlogPostSchema) } },
    },
    async (request, reply) => {
      const {
        title,
        slug,
        excerpt,
        contentHtml,
        coverImageUrl,
        status,
        categoryId,
        seoTitle,
        seoDescription,
        ogTitle,
        ogImageUrl,
        canonicalUrl,
        noIndex,
        translations,
      } = request.body;

      const post = await app.prisma.blogPost.create({
        data: {
          title,
          slug: slug ? slugify(slug) : slugify(title),
          excerpt,
          contentHtml: contentHtml ?? "",
          coverImageUrl,
          status: status ?? "DRAFT",
          categoryId: categoryId ?? undefined,
          authorId: request.user!.id,
          seoTitle,
          seoDescription,
          ogTitle,
          ogImageUrl,
          canonicalUrl,
          noIndex,
          translations: (translations ?? {}) as Prisma.InputJsonValue,
          publishedAt: status === "PUBLISHED" ? new Date() : null,
        },
        include: WITH_CATEGORY,
      });

      return reply.code(201).send(ok(toBlogPostDto(post)));
    }
  );

  server.get(
    "/:postId",
    { schema: { params: PostIdParamSchema, response: { 200: ApiSuccessSchema(BlogPostSchema) } } },
    async (request, reply) => {
      const post = await app.prisma.blogPost.findUnique({
        where: { id: request.params.postId },
        include: WITH_CATEGORY,
      });
      if (!post) throw new NotFoundError("Yazı bulunamadı.");
      return reply.send(ok(toBlogPostDto(post)));
    }
  );

  server.patch(
    "/:postId",
    {
      preHandler: requireSiteRole("ADMIN", "EDITOR"),
      schema: {
        params: PostIdParamSchema,
        body: UpdateBlogPostRequestSchema,
        response: { 200: ApiSuccessSchema(BlogPostSchema) },
      },
    },
    async (request, reply) => {
      const existing = await app.prisma.blogPost.findUnique({ where: { id: request.params.postId } });
      if (!existing) throw new NotFoundError("Yazı bulunamadı.");

      await snapshotBeforeUpdate(app, "BLOG_POST", existing.id, toBlogPostSnapshot(existing), request.user!.id);

      const { slug, translations, ...rest } = request.body;

      const mergedTranslations =
        translations !== undefined
          ? {
              ...((existing.translations as Record<string, Record<string, unknown>>) ?? {}),
              ...Object.fromEntries(
                Object.entries(translations).map(([locale, fields]) => [
                  locale,
                  { ...(((existing.translations as Record<string, Record<string, unknown>>) ?? {})[locale] ?? {}), ...fields },
                ])
              ),
            }
          : undefined;

      const post = await app.prisma.blogPost.update({
        where: { id: request.params.postId },
        data: {
          ...rest,
          ...(slug !== undefined ? { slug: slugify(slug) } : {}),
          ...(mergedTranslations !== undefined ? { translations: mergedTranslations as Prisma.InputJsonValue } : {}),
          ...(rest.status === "PUBLISHED" && !existing.publishedAt ? { publishedAt: new Date() } : {}),
        },
        include: WITH_CATEGORY,
      });

      return reply.send(ok(toBlogPostDto(post)));
    }
  );

  server.delete(
    "/:postId",
    {
      preHandler: requireSiteRole("ADMIN"),
      schema: { params: PostIdParamSchema, response: { 204: z.undefined() } },
    },
    async (request, reply) => {
      await app.prisma.blogPost.delete({ where: { id: request.params.postId } }).catch(() => {
        throw new NotFoundError("Yazı bulunamadı.");
      });
      return reply.code(204).send();
    }
  );

  // §10.1 İçerik Sürüm Kontrolü — yetki eşiği yazı düzenleme ile aynı (ADMIN+EDITOR).
  server.get(
    "/:postId/revisions",
    {
      preHandler: requireSiteRole("ADMIN", "EDITOR"),
      schema: {
        params: PostIdParamSchema,
        querystring: CursorQuerySchema,
        response: { 200: ApiSuccessSchema(z.array(ContentRevisionSummarySchema)) },
      },
    },
    async (request, reply) => {
      const { cursor, limit } = request.query;
      const cursorSeq = parseCursor(cursor);

      const rows = await app.prisma.contentRevision.findMany({
        where: {
          entityType: "BLOG_POST",
          entityId: request.params.postId,
          ...(cursorSeq ? { seq: { lt: cursorSeq } } : {}),
        },
        orderBy: { seq: "desc" },
        take: limit,
      });

      const nextCursor = rows.length === limit ? encodeCursor(rows[rows.length - 1]!.seq) : null;

      return reply.send(ok(rows.map(toContentRevisionSummaryDto), { nextCursor }));
    }
  );

  server.get(
    "/:postId/revisions/:revisionId",
    {
      preHandler: requireSiteRole("ADMIN", "EDITOR"),
      schema: { params: PostRevisionIdParamSchema, response: { 200: ApiSuccessSchema(ContentRevisionSchema) } },
    },
    async (request, reply) => {
      const revision = await app.prisma.contentRevision.findUnique({ where: { id: request.params.revisionId } });
      if (!revision || revision.entityType !== "BLOG_POST" || revision.entityId !== request.params.postId) {
        throw new NotFoundError("Revizyon bulunamadı.");
      }
      return reply.send(ok(toContentRevisionDto(revision)));
    }
  );

  server.post(
    "/:postId/revisions/:revisionId/restore",
    {
      preHandler: requireSiteRole("ADMIN", "EDITOR"),
      schema: { params: PostRevisionIdParamSchema, response: { 200: ApiSuccessSchema(BlogPostSchema) } },
    },
    async (request, reply) => {
      const existing = await app.prisma.blogPost.findUnique({ where: { id: request.params.postId } });
      if (!existing) throw new NotFoundError("Yazı bulunamadı.");

      const revision = await app.prisma.contentRevision.findUnique({ where: { id: request.params.revisionId } });
      if (!revision || revision.entityType !== "BLOG_POST" || revision.entityId !== request.params.postId) {
        throw new NotFoundError("Revizyon bulunamadı.");
      }

      // Geri dönüş de geri alınabilir olsun diye önce mevcut state'i yeni bir revizyon olarak kaydet.
      await snapshotBeforeUpdate(app, "BLOG_POST", existing.id, toBlogPostSnapshot(existing), request.user!.id);

      const snapshot = revision.snapshot as {
        title: string;
        slug: string;
        excerpt: string | null;
        contentHtml: string;
        coverImageUrl: string | null;
        categoryId: string | null;
        seoTitle: string | null;
        seoDescription: string | null;
        ogTitle: string | null;
        ogImageUrl: string | null;
        canonicalUrl: string | null;
        noIndex: boolean;
        translations: unknown;
      };

      const post = await app.prisma.blogPost.update({
        where: { id: request.params.postId },
        data: {
          title: snapshot.title,
          slug: snapshot.slug,
          excerpt: snapshot.excerpt,
          contentHtml: snapshot.contentHtml,
          coverImageUrl: snapshot.coverImageUrl,
          categoryId: snapshot.categoryId,
          seoTitle: snapshot.seoTitle,
          seoDescription: snapshot.seoDescription,
          ogTitle: snapshot.ogTitle,
          ogImageUrl: snapshot.ogImageUrl,
          canonicalUrl: snapshot.canonicalUrl,
          noIndex: snapshot.noIndex,
          translations: (snapshot.translations ?? {}) as Prisma.InputJsonValue,
        },
        include: WITH_CATEGORY,
      });

      return reply.send(ok(toBlogPostDto(post)));
    }
  );
}

/** `/admin/blog/categories` prefix'i altında bağlanır — authenticated. */
export async function adminBlogCategoriesRoutes(app: FastifyInstance) {
  const server = app.withTypeProvider<ZodTypeProvider>();
  server.addHook("preHandler", authenticate);

  server.get(
    "/",
    { schema: { response: { 200: ApiSuccessSchema(z.array(BlogCategorySchema)) } } },
    async (_request, reply) => {
      const rows = await app.prisma.blogCategory.findMany({ orderBy: { seq: "asc" } });
      return reply.send(ok(rows.map(toBlogCategoryDto)));
    }
  );

  server.post(
    "/",
    {
      preHandler: requireSiteRole("ADMIN", "EDITOR"),
      schema: { body: CreateBlogCategoryRequestSchema, response: { 201: ApiSuccessSchema(BlogCategorySchema) } },
    },
    async (request, reply) => {
      const category = await app.prisma.blogCategory.create({
        data: {
          name: request.body.name,
          slug: request.body.slug ? slugify(request.body.slug) : slugify(request.body.name),
        },
      });
      return reply.code(201).send(ok(toBlogCategoryDto(category)));
    }
  );

  server.patch(
    "/:categoryId",
    {
      preHandler: requireSiteRole("ADMIN", "EDITOR"),
      schema: {
        params: CategoryIdParamSchema,
        body: UpdateBlogCategoryRequestSchema,
        response: { 200: ApiSuccessSchema(BlogCategorySchema) },
      },
    },
    async (request, reply) => {
      const { slug, ...rest } = request.body;
      const category = await app.prisma.blogCategory
        .update({
          where: { id: request.params.categoryId },
          data: { ...rest, ...(slug !== undefined ? { slug: slugify(slug) } : {}) },
        })
        .catch(() => {
          throw new NotFoundError("Kategori bulunamadı.");
        });
      return reply.send(ok(toBlogCategoryDto(category)));
    }
  );

  server.delete(
    "/:categoryId",
    {
      preHandler: requireSiteRole("ADMIN"),
      schema: { params: CategoryIdParamSchema, response: { 204: z.undefined() } },
    },
    async (request, reply) => {
      await app.prisma.blogCategory.delete({ where: { id: request.params.categoryId } }).catch(() => {
        throw new NotFoundError("Kategori bulunamadı.");
      });
      return reply.code(204).send();
    }
  );
}

/** `/blog` prefix'i altında bağlanır — herkese açık, yalnızca yayınlanmış yazılar. */
export async function publicBlogRoutes(app: FastifyInstance) {
  const server = app.withTypeProvider<ZodTypeProvider>();

  server.get(
    "/",
    { schema: { querystring: CursorQuerySchema, response: { 200: ApiSuccessSchema(z.array(BlogPostSchema)) } } },
    async (request, reply) => {
      const { cursor, limit } = request.query;
      const cursorSeq = parseCursor(cursor);

      const rows = await app.prisma.blogPost.findMany({
        where: { status: "PUBLISHED", ...(cursorSeq ? { seq: { gt: cursorSeq } } : {}) },
        orderBy: { seq: "asc" },
        take: limit,
        include: WITH_CATEGORY,
      });

      return reply.send(ok(rows.map(toBlogPostDto), buildPageMeta(rows, limit)));
    }
  );

  server.get(
    "/:slug",
    {
      schema: { params: PostSlugParamSchema, querystring: LocaleQuerySchema, response: { 200: ApiSuccessSchema(BlogPostSchema) } },
    },
    async (request, reply) => {
      const post = await app.prisma.blogPost.findFirst({
        where: { slug: request.params.slug, status: "PUBLISHED" },
        include: WITH_CATEGORY,
      });
      if (!post) throw new NotFoundError("Yazı bulunamadı.");
      return reply.send(ok(toBlogPostDto(applyLocale(post, request.query.locale))));
    }
  );

  server.post(
    "/:slug/view",
    { schema: { params: PostSlugParamSchema, response: { 204: z.undefined() } } },
    async (request, reply) => {
      const post = await app.prisma.blogPost.findFirst({
        where: { slug: request.params.slug, status: "PUBLISHED" },
        select: { id: true },
      });
      if (!post) throw new NotFoundError("Yazı bulunamadı.");

      const deviceType = detectDeviceType(request.headers["user-agent"]);
      const country = detectCountry(request.ip);
      touchVisitor(request.ip, request.headers["user-agent"]);

      const date = startOfUtcDay();
      await app.prisma.$transaction([
        app.prisma.pageView.upsert({
          where: { postId_date_deviceType_country: { postId: post.id, date, deviceType, country } },
          create: { postId: post.id, date, deviceType, country, count: 1 },
          update: { count: { increment: 1 } },
        }),
        app.prisma.blogPost.update({ where: { id: post.id }, data: { viewCount: { increment: 1 } } }),
      ]);

      return reply.code(204).send();
    }
  );
}