# docs/architecture/openapi.yaml
openapi: 3.0.3
info:
  title: SaaS Platform API
  version: "1.0"
  description: >
    Frontend (Next.js) ve backend (Node.js/Fastify) ajanlarının uyacağı API
    kontratı. Kaynak doğruluk: bu dosya + ../shared-types.ts + ARCHITECTURE.md.
servers:
  - url: https://api.example.com/api/v1
    description: Production
  - url: http://localhost:4000/api/v1
    description: Local
...
security:
  - bearerAuth: []
paths:
...
  /admin/logs:
    get:
      tags: [Logs]
      summary: Audit log kayıtlarını listele (yalnızca ADMIN)
      description: >
        SAPMA: diğer tüm sayfalanan listelerin aksine en yeni kayıt önce gelir
        (`orderBy seq desc` + `seq < cursor`), diğer uçlarda `orderBy seq asc` + `seq > cursor`.
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: status
          in: query
          required: false
          schema: { type: string, enum: [SUCCESS, FAILURE, FORBIDDEN] }
        - name: action
          in: query
          required: false
          description: Öneke göre filtreler (ör. "user." → user.create, user.role_change, ...).
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccessEnvelope"
                  - properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/AuditLog" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /admin/blog/{postId}/revisions:
    parameters:
      - $ref: "#/components/parameters/PostId"
    get:
      tags: [ContentRevisions]
      summary: Yazının revizyon geçmişini listele (ADMIN/EDITOR)
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccessEnvelope"
                  - properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/ContentRevisionSummary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/blog/{postId}/revisions/{revisionId}:
    parameters:
      - $ref: "#/components/parameters/PostId"
      - $ref: "#/components/parameters/RevisionId"
    get:
      tags: [ContentRevisions]
      summary: Bir revizyonun tam snapshot'ını getir (ADMIN/EDITOR)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccessEnvelope"
                  - properties:
                      data: { $ref: "#/components/schemas/ContentRevision" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /admin/blog/{postId}/revisions/{revisionId}/restore:
    parameters:
      - $ref: "#/components/parameters/PostId"
      - $ref: "#/components/parameters/RevisionId"
    post:
      tags: [ContentRevisions]
      summary: Yazıyı bu revizyondaki hale geri döndür (ADMIN/EDITOR)
      description: >
        Geri dönüş de geri alınabilsin diye önce mevcut state yeni bir revizyon
        olarak kaydedilir, ardından hedef revizyonun snapshot'ı uygulanır.
      responses:
        "200":
          description: Yazı geri yüklendi
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiSuccessEnvelope"
                  - properties:
                      data: { $ref: "#/components/schemas/BlogPost" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, default: 20, maximum: 100 }
...
    PostId:
      name: postId
      in: path
      required: true
      schema: { type: string, format: uuid }
    RevisionId:
      name: revisionId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Unauthorized:
      description: Kimlik doğrulama yok/geçersiz
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiErrorEnvelope" }
    Forbidden:
      description: Yetki yok
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiErrorEnvelope" }
    NotFound:
      description: Kaynak bulunamadı
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiErrorEnvelope" }
...
  schemas:
    ApiSuccessEnvelope:
      type: object
      required: [data]
      properties:
        data: {}
        meta:
          type: object
          properties:
            nextCursor: { type: string, nullable: true }
    ApiErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [UNAUTHORIZED, FORBIDDEN, NOT_FOUND, VALIDATION_ERROR, CONFLICT, RATE_LIMITED, INTERNAL_ERROR]
            message: { type: string }
            details:
              type: object
              additionalProperties:
                type: array
                items: { type: string }
...
    AuditLog:
      type: object
      properties:
        id: { type: string, format: uuid }
        actorId: { type: string, format: uuid, nullable: true }
        actorEmail: { type: string, nullable: true }
        action: { type: string, example: "user.role_change" }
        status: { type: string, enum: [SUCCESS, FAILURE, FORBIDDEN] }
        targetType: { type: string, nullable: true }
        targetId: { type: string, nullable: true }
        metadata:
          type: object
          nullable: true
          additionalProperties: true
        ipAddress: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
...
    BlogCategory:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        createdAt: { type: string, format: date-time }
...
    BlogPost:
      description: >
        `/admin/blog` ve `/blog` uçlarında dönen yazı DTO'su. `translations.EN` yalnızca
        override'ları taşır (bkz. ARCHITECTURE.md §10.5).
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        slug: { type: string }
        excerpt: { type: string, nullable: true }
        contentHtml: { type: string }
        coverImageUrl: { type: string, nullable: true }
        status: { type: string, enum: [DRAFT, PUBLISHED] }
        category:
          allOf:
            - $ref: "#/components/schemas/BlogCategory"
          nullable: true
        seoTitle: { type: string, nullable: true }
        seoDescription: { type: string, nullable: true }
        ogTitle: { type: string, nullable: true }
        ogImageUrl: { type: string, nullable: true }
        canonicalUrl: { type: string, nullable: true }
        noIndex: { type: boolean }
        translations:
          type: object
          description: "ör. { \"EN\": { \"title\": \"...\", \"contentHtml\": \"...\" } } — bkz. shared-types.ts::BlogPostTranslations"
          additionalProperties:
            type: object
            additionalProperties: true
        publishedAt: { type: string, format: date-time, nullable: true }
        viewCount: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ContentEntityType:
      type: string
      enum: [PAGE, BLOG_POST]
    ContentRevisionSummary:
      description: >
        Revizyon liste görünümü — `snapshot` alanını İÇERMEZ (bkz. ARCHITECTURE.md §10.1).
      type: object
      properties:
        id: { type: string, format: uuid }
        editedById: { type: string, format: uuid, nullable: true }
        editedByName: { type: string, description: "silinen kullanıcı için okunabilir ad snapshot'ı" }
        createdAt: { type: string, format: date-time }
    ContentRevision:
      allOf:
        - $ref: "#/components/schemas/ContentRevisionSummary"
        - type: object
          properties:
            entityType: { $ref: "#/components/schemas/ContentEntityType" }
            entityId: { type: string }
            snapshot:
              type: object
              description: >
                Güncellemeden HEMEN ÖNCEKİ tam alan seti — Page için `{ title, slug,
                blocks, seoTitle, seoDescription, ogTitle, ogImageUrl, canonicalUrl,
                noIndex, translations }`; BlogPost için `{ title, slug, excerpt,
                contentHtml, coverImageUrl, categoryId, seoTitle, seoDescription,
                ogTitle, ogImageUrl, canonicalUrl, noIndex, translations }`.
              additionalProperties: true
...